Turn doc comments in sensirion_i2c.h into normal comments.
Functions are documented multiple times. Once in the header file
and once in each implementation (sample implemenations and
implementation templates). Doxygen cannot handle this when the
documentations differ slightly and produces duplicate output.
For exapmle when the header file contains comments for hardware
and software i2c and an implemenation template does not.
As a admittedly hacky fix we turn the doc comments in the header
file into normal comments (/** -> /*). This way doxygen does not
pick up the comments in the header file.